+++ adapters/Pangle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next Version
 - Fix an issue where GADPangleNetworkExtras did not conform to GADAdNetworkExtras protocol.
+- Implement adDidShowFail protocol method in interstitial, rewarded, and app open ads.
 
 #### [Version 5.3.1.2.0](https://dl.google.com/googleadmobadssdk/mediation/ios/pangle/PangleAdapter-5.3.1.2.0.zip)
 - Verified compatibility with Pangle SDK version 5.3.1.2.